Create MatrixGenerator.py
Uses Pygame so make sure you have installed, needs an extra line above the data as too lazy to fix and it works fine, x and y at the top you can also play around with as the get_screen.width() and height wasnt working in the random numbers below. This is the basic version with no implemented features to make it look better either so... yes.